+++ src/wirestate-react/provision/use-scope.test.tsx
@@ -1,0 +1,27 @@
+import { render } from "@testing-library/react";
+import { Container, createIocContainer, WireScope } from "@wirestate/core";
+
+import { withIocProvider } from "../test-utils/with-ioc-provider";
+import { Optional } from "../types/general";
+
+import { useScope } from "./use-scope";
+
+describe("useScope", () => {
+  it("should return current container scope", () => {
+    const container: Container = createIocContainer();
+    let scope: Optional<WireScope> = null as Optional<WireScope>;
+
+    function TestComponent() {
+      scope = useScope();
+
+      return null;
+    }
+
+    render(withIocProvider(<TestComponent />, container));
+
+    expect(scope).not.toBeNull();
+    expect(scope).toBeInstanceOf(WireScope);
+    expect(scope).not.toBe(container.get(WireScope));
+    expect(scope?.getContainer()).toBe(container);
+  });
+});